refactor: extract admin skill health projection
